+++ packages/albus-cli/src/actions/request/remove.ts
@@ -30,7 +30,9 @@ import log from 'loglevel'
 import { exploreTransaction } from '@/utils'
 import { useContext } from '@/context'
 
-type Opts = {}
+type Opts = {
+  // ...
+}
 
 export async function remove(addr: string, _opts: Opts) {
   const { client } = useContext()